refactor(Gamer): Playable 인터페이스를 추상 클래스로 변경 및 공통 부분 통일
