--- src/main/java/blackjack/model/participant/Playable.java
@@ -1,14 +1,0 @@
-package blackjack.model.participant;
-
-import blackjack.model.deck.Card;
-import java.util.List;
-
-public interface Playable {
-
-    void receiveCard(final Card card);
-    List<Card> openCards();
-    int getScore();
-    int getCardCount();
-    boolean canHit();
-    boolean isBust();
-}
